refactor(url): extract and modularize url extraction logic into a new base module
This change separates the URL extraction and exploration strategy logic into a dedicated `base.py` module. The update improves modularity and reusability by removing duplicate code and simplifying imports in other related modules.
